+++ src/main/java/com/atlauncher/data/Instance.java
@@ -40,10 +40,12 @@ import com.atlauncher.Gsons;
 import com.atlauncher.LogManager;
 import com.atlauncher.data.openmods.OpenEyeReportResponse;
 import com.atlauncher.gui.dialogs.ProgressDialog;
-import com.atlauncher.mclauncher.LegacyMCLauncher;
 import com.atlauncher.mclauncher.MCLauncher;
 import com.atlauncher.utils.HTMLUtils;
 import com.atlauncher.utils.Utils;
+
+import io.sentry.Sentry;
+import io.sentry.event.BreadcrumbBuilder;
 
 import com.atlauncher.exceptions.InvalidMinecraftVersion;
 
@@ -1415,6 +1417,8 @@ public class Instance implements Cloneable {
                             }
                             LogManager.minecraft(line);
                         }
+                        Sentry.getContext().recordBreadcrumb(
+                                new BreadcrumbBuilder().setMessage("Launched instance closed").build());
                         App.settings.hideKillMinecraft();
                         if (App.settings.getParent() != null && App.settings.keepLauncherOpen()) {
                             App.settings.getParent().setVisible(true);