+++ CHANGELOG.md
@@ -10,3 +10,4 @@
 -   Make the window size a little bit bigger
 -   Change references to NodeCraft to Nodecraft
 -   Don't download resources when creating a server
+-   Add Sentry error logging to help debug issues